# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Сваггер сгенерирован из кода, поэтому может содержать неточности. По мере возможности они будут описаны далее\nВсе даты передаются как строки\nБольшая часть рутов не возвращает вложенные объекты",
        "title": "GenMilk API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "genmilk.ru",
    "basePath": "/api",
    "paths": {
...
        "/user/whoami": {
            "get": {
                "description": "Рут вернет данные о пользователе из БД",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Получить информацию о пользователе",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
...
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}